+++ test/dummy/config/routes.rb
@@ -2,5 +2,6 @@
 
 Rails.application.routes.draw do
   get "/preview", to: "previews#new"
+  get "/preview/hostile", to: "previews#hostile"
   post "/preview", to: "previews#create"
 end